More changes due to refactor
